requirements.txt
# HighwayRetroAI — Python Dependencies
# Pin versions for reproducibility

# Core ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.2.0
efficientnet-pytorch>=0.7.1
onnx>=1.14.0
onnxruntime>=1.16.0

# Vision
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0

# IoT / Hardware
smbus2>=0.4.2
pyserial>=3.5

# Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Dashboard
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pynmea2>=1.19.0
requests>=2.31.0

# Training extras
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0

# On Jetson Xavier NX:
# 1. Install PyTorch from NVIDIA wheel (not pip):
#    https://developer.download.nvidia.com/compute/redist/jp/
# 2. torch-tensorrt: pip install torch-tensorrt --find-links ...
# 3. grovepi: pip install grovepi  (if using Grove Pi hat)
# 4. For FLIR Lepton: pip install pylepton
# 5. For DHT11: pip install adafruit-circuitpython-dht Adafruit-DHT
# 6. For Jetson GPIO: pip install Jetson.GPIO
